Edit README.md
Removed the section on usage as that section pertains more to
projects intended for developers and a usage section might chronicle
use cases. For Yodel the Quick Start section is sufficient. Commenting
out the requirements and develoment sections because this may be
unnecessary for the intended audience of this readme--namely the product
user. Information relating to development could be moved to
CONTRIBUTING.md.